Updates to fix the focus on scroll position and add the tooltip directive.
Adds a tooltip directive so we can easily modify the view and styles
associated with our tooltips.  Adds a task to the gulp configuration for
"html2js" that will convert the HTML templates into JS that loads the
HTML into the template cache so the directive can load the template from
there.

Tooltip directive adds an element as a sibling for whatever element it
is applied to so the positioning can be achieved.  Currently the
calculations have a few hard coded values that need to be elevated to
constants (namely 10 for the size of the arrow).  Ideally we would also
have a LESS or SASS compilation step in the build process for the utils
so we can avoid the more verbose CSS used here.