Modified how the position is calculated to make it more accurate.
Also pretty big change here is the tooltips are being added to the body
instead of as a sibling to the element they modify.  When appending
after the current element there were issues with how the width is
calculated.  The new method I was able to more precisely line up the
tooltip with the element and allows the tool tip to resize based on it's
child contents or have a max width set to allow control of how the text
wraps.