Last changes to get everything building for the sake of loading into the main-app
Had to add some build steps for copying over the css so it can be loaded
into the main-app.  Did some actual testing with loading this up in the
main app and testing the new tooltips so I know this build process is
good enough at this point.  We still need to manually load the css
separately from the townsquared.css from how I changed the main-app but
this may be resolved in the future.  For now this is functional for
loading the HTML templates and CSS from the module.