Updates to get drop downs working in a basic way at least.
Still needs some finishing up with regard to keyboard event handling and
customization of list items to allow elements in the list that are not
selectable and for handling default selection properly.  Does the basics
of open up allow a user to click an element, has custom templates for
the list items and a template for the placeholder/chosen value display.
Intelligently opens up or down depending on the position of the
element on screen, if more than half way down opens upward and reverses
the list items, otherwise opens downwards.